+++ src/math/arrayCumulativeSum.ts
@@ -1,8 +1,9 @@
-export function arrayCumulativeSum(array: Float64Array): Float64Array {
+export function arrayCumulativeSum(array: Float32Array): Float64Array {
   if (array.length < 1) {
     throw new Error('cumulative sum expects values in the given array');
   }
 
+  // TODO: this needs to stay f64 for proper bass ratio calculation, why?
   const sum = new Float64Array(array.length);
   // eslint-disable-next-line prefer-destructuring
   sum[0] = array[0];
