+++ src/math/arrayFilledWithZeros.ts
@@ -1,3 +1,3 @@
-export function arrayFilledWithZeros(length: number): Float64Array {
-  return new Float64Array(length);
+export function arrayFilledWithZeros(length: number): Float32Array {
+  return new Float32Array(length);
 }
