+++ test/starttime.test.ts
@@ -6,22 +6,22 @@ import {
 } from '../src/starttime';
 
 it('trims samples to correct starttime', () => {
-  const samples = new Float64Array([
+  const samples = new Float32Array([
     0, 0, 0.001, 0.001, 0.01, 0.02, 0.03, 0.04, 1, 0.001, 0, 0,
   ]);
 
   const result = correctStarttimeMonaural(samples);
 
   expect(result).deep.equal(
-    new Float64Array([0.02, 0.03, 0.04, 1, 0.001, 0, 0])
+    new Float32Array([0.02, 0.03, 0.04, 1, 0.001, 0, 0])
   );
 });
 
 it('trims binaural samples to correct starttime', () => {
-  const leftSamples = new Float64Array([
+  const leftSamples = new Float32Array([
     0, 0, 0.001, 0.001, 0.01, 0.02, 0.03, 0.04, 1, 0.001, 0,
   ]);
-  const rightSamples = new Float64Array([
+  const rightSamples = new Float32Array([
     0, 0.001, 0.001, 0.01, 0.02, 0.03, 0.04, 1, 0.001, 0, 0,
   ]);
 
@@ -30,9 +30,9 @@ it('trims binaural samples to correct starttime', () => {
   );
 
   expect(result.leftSamples).deep.equal(
-    new Float64Array([0.01, 0.02, 0.03, 0.04, 1, 0.001, 0])
+    new Float32Array([0.01, 0.02, 0.03, 0.04, 1, 0.001, 0])
   );
   expect(result.rightSamples).deep.equal(
-    new Float64Array([0.02, 0.03, 0.04, 1, 0.001, 0, 0])
+    new Float32Array([0.02, 0.03, 0.04, 1, 0.001, 0, 0])
   );
 });
